add timer; add unit tests for it;

+++ src/BeeInstant.NetSDK/Abstractions/IRecorder.cs
@@ -2,6 +2,6 @@ namespace BeeInstant.NetSDK.Abstractions
 {
     public interface IRecorder : IMetric<IRecorder>
     {
-        void Record(double value, Unit unit);
+        void Record(decimal value, Unit unit);
     }
 }
